# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "hoydigo API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/admin/position": {
            "post": {
                "tags": [
                    "Admin - Position"
                ],
                "summary": "Create a new position",
                "description": "Create a new position",
                "operationId": "createPosition",
                "requestBody": {
                    "description": "Fields for the new position",
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "Id for the new position",
                                        "type": "string",
                                        "format": "text",
                                        "example": "MYPOS"
                                    },
                                    "country_id": {
                                        "description": "Country id where the position is",
                                        "type": "string",
                                        "example": "COL"
                                    },
                                    "state_id": {
                                        "description": "The state id should be in the country indicated",
                                        "type": "integer",
                                        "example": "5"
                                    },
                                    "city_id": {
                                        "description": "The city id should be in the state and country indicated",
                                        "type": "integer",
                                        "example": "1017"
                                    },
                                    "sector": {
                                        "description": "Sector where the position works",
                                        "type": "string",
                                        "example": "Politico"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Senador de la Republica"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descripción del cargo de senador"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Position created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Position created successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "format": "text",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "User without permission to the endpoint",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "format": "text",
                                            "example": "Permission denied."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "example": "{'field':['Error message']}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "name": "Bearer Token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}